Update tournament selection strategy

+++ ea/genotype.py
@@ -9,7 +9,7 @@ class Genotype:
         else:
             self.bulbs = set([])
 
-        self.fitness = int(config.settings['arbitrary_large_number'])
+        self.fitness = float(config.settings['arbitrary_large_number'])
         self.shine_ratio = 0.0
         self.black_cell_constraints_violated = 0
         self.bulb_shine_constraints_violated = 0
